Renamed to correspond to the use case name.
