+++ src/main/java/data_access/summarization/InMemorySummarizationDAO.java
@@ -1,4 +1,4 @@
-package data_access.outfit;
+package data_access.summarization;
 
 /**
  * In memory implementation of the Outfit data access object.
