+++ src/main/java/data_access/summarization/SummarizationDAO.java
@@ -1,4 +1,4 @@
-package data_access.outfit;
+package data_access.summarization;
 
 // TODO: Implement the OutfitDataAccessObject class [Kush]
 public class SummarizationDAO {
